+++ segtypes/common/rodata.py
@@ -1,4 +1,6 @@
+from typing import Optional, Set, Tuple
 import spimdisasm
+from segtypes.segment import Segment
 from util import log, options, symbols
 
 from segtypes.common.data import CommonSegData
@@ -7,6 +9,27 @@ from segtypes.common.data import CommonSegData
 class CommonSegRodata(CommonSegData):
     def get_linker_section(self) -> str:
         return ".rodata"
+
+    def get_possible_text_subsegment_for_symbol(
+        self, rodata_sym: spimdisasm.mips.symbols.SymbolBase
+    ) -> Optional[Tuple[Segment, spimdisasm.common.ContextSymbol]]:
+        # Check if this rodata segment does not have a corresponding code file, try to look for one
+
+        if self.sibling is not None or not options.opts.pair_rodata_to_text:
+            return None
+
+        if not rodata_sym.shouldMigrate():
+            return None
+
+        if len(rodata_sym.contextSym.referenceFunctions) != 1:
+            return None
+
+        func = list(rodata_sym.contextSym.referenceFunctions)[0]
+        text_segment = self.parent.get_subsegment_for_ram(func.vram)
+
+        if text_segment is None or not text_segment.is_text():
+            return None
+        return text_segment, func
 
     def disassemble_data(self, rom_bytes):
         if not isinstance(self.rom_start, int):
@@ -49,10 +72,24 @@ class CommonSegRodata(CommonSegData):
         self.spim_section.analyze()
         self.spim_section.setCommentOffset(self.rom_start)
 
+        possible_text_segments: Set[Segment] = set()
+
         for symbol in self.spim_section.symbolList:
             symbols.create_symbol_from_spim_symbol(
                 self.get_most_parent(), symbol.contextSym
             )
+
+            possible_text = self.get_possible_text_subsegment_for_symbol(symbol)
+            if possible_text is not None:
+                text_segment, refenceeFunction = possible_text
+                if text_segment not in possible_text_segments:
+                    print(
+                        f"\nRodata segment '{self.name}' may belong to the text segment '{text_segment.name}'"
+                    )
+                    print(
+                        f"    Based on the usage from the function {refenceeFunction.getName()} to the symbol {symbol.getName()}"
+                    )
+                    possible_text_segments.add(text_segment)
 
     def split(self, rom_bytes: bytes):
         # Disassemble the file itself